feat: move data types into respective modules
