+++ users/src/data/mod.rs
@@ -1,0 +1,2 @@
+pub mod role;
+pub mod user;
